requirements.txt
# Core Framework
streamlit>=1.28.0

# AI Providers
anthropic>=0.7.0
openai>=1.45.0

# API Interactions
requests>=2.31.0
PyGithub>=2.1.1

# Data Validation
pydantic>=2.4.0
pydantic-settings>=2.0.3

# File Processing
pyyaml>=6.0.1
Jinja2>=3.1.2

# Environment Variables
python-dotenv>=1.0.0

# Utilities
pandas>=2.1.0
